Removed gif support and simplified library
