+++ composer.json
@@ -14,14 +14,8 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^8.4",
-        "sybio/gif-creator": "^1.0",
-        "sybio/gif-frame-extractor": "^1.0",
         "symfony/filesystem": "^4.3",
         "vimeo/psalm": "^3.11"
-    },
-    "suggest": {
-        "sybio/gif-creator": "Install for animated GIF support",
-        "sybio/gif-frame-extractor": "Install for animated GIF support"
     },
     "autoload": {
         "psr-4": {
