+++ src/Exception/UnsupportedExtensionException.php
@@ -5,10 +5,10 @@ namespace Yannickl88\Image\Exception;
 
 class UnsupportedExtensionException extends ImageException
 {
-    public function __construct(string $extension, array $suported, \Throwable $previous = null)
+    public function __construct(string $extension, array $supported, \Throwable $previous = null)
     {
         parent::__construct(
-            sprintf('Unsupported file format "%s", supported are: "%s".', $extension, implode('", "', $suported)),
+            sprintf('Unsupported file format "%s", supported are: "%s".', $extension, implode('", "', $supported)),
             0,
             $previous
         );
