+++ tests/StaticImageTest.php
@@ -3,10 +3,12 @@ namespace Yannickl88\Image;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
+use Yannickl88\Image\Exception\FileNotFoundException;
 use Yannickl88\Image\Exception\ImageException;
+use Yannickl88\Image\Exception\UnsupportedExtensionException;
 
 /**
- * @covers \Yannickl88\Image\StaticImage
+ * @covers \Yannickl88\Image\Image
  */
 class StaticImageTest extends TestCase
 {
@@ -17,20 +19,138 @@ class StaticImageTest extends TestCase
         self::$output_dir = __DIR__ . '/' . uniqid('out_', false);
     }
 
+    public function allFilesProvider()
+    {
+        return [
+            [Image::class, __DIR__ . '/resources/image1.gif'],
+            [Image::class, __DIR__ . '/resources/image2.png'],
+            [Image::class, __DIR__ . '/resources/image3.jpg'],
+            [Image::class, __DIR__ . '/resources/image4.jpeg'],
+            [Image::class, __DIR__ . '/resources/image5.gif'],
+        ];
+    }
+
+    public function orientationProvider()
+    {
+        return [
+            [Image::ORIENTATION_LANDSCAPE, __DIR__ . '/resources/image1.gif'],
+            [Image::ORIENTATION_PORTRAIT, __DIR__ . '/resources/image2.png'],
+            [Image::ORIENTATION_LANDSCAPE, __DIR__ . '/resources/image3.jpg'],
+            [Image::ORIENTATION_LANDSCAPE, __DIR__ . '/resources/image4.jpeg'],
+            [Image::ORIENTATION_LANDSCAPE, __DIR__ . '/resources/image5.gif'],
+        ];
+    }
+
+    public function fitProvider(): array
+    {
+        return [
+            [[0, 0, 100, 69], false, __DIR__ . '/resources/image1.gif'],
+            [[0, 0, 100, 69], true, __DIR__ . '/resources/image1.gif'],
+            [[0, 0, 80, 100], false, __DIR__ . '/resources/image2.png'],
+            [[0, 0, 80, 100], true, __DIR__ . '/resources/image2.png'],
+            [[0, 0, 100, 67], false, __DIR__ . '/resources/image3.jpg'],
+            [[0, 0, 100, 67], true, __DIR__ . '/resources/image3.jpg'],
+            [[0, 0, 100, 67], false, __DIR__ . '/resources/image4.jpeg'],
+            [[0, 0, 100, 67], true, __DIR__ . '/resources/image4.jpeg'],
+            [[0, 0, 100, 67], false, __DIR__ . '/resources/image5.gif'],
+            [[0, 0, 100, 67], true, __DIR__ . '/resources/image5.gif'],
+            [[0, 0, 60, 75], false, __DIR__ . '/resources/image7.png'],
+            [[0, 0, 80, 100], true, __DIR__ . '/resources/image7.png'],
+        ];
+    }
+
     protected function tearDown(): void
     {
         $fs = new Filesystem();
         $fs->remove(self::$output_dir);
     }
 
+    /**
+     * @dataProvider allFilesProvider
+     */
+    public function testFromFile(string $expected_class, string $filename): void
+    {
+        self::assertInstanceOf($expected_class, Image::fromFile($filename));
+    }
+
+    public function testFromBadFileName(): void
+    {
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('File name does not contain a . extension (example.jpg)');
+
+        Image::fromFile(__DIR__ . '/resources/foobar');
+    }
+
+    public function testFromFileNotExists(): void
+    {
+        $this->expectException(FileNotFoundException::class);
+
+        Image::fromFile('foobar.gif');
+    }
+
+    public function testFromFileUnsupported(): void
+    {
+        $this->expectException(UnsupportedExtensionException::class);
+
+        Image::fromFile(__DIR__ . '/resources/image6.bmp');
+    }
+
+    public function testFromFileCorrupt(): void
+    {
+        $this->expectException(ImageException::class);
+
+        Image::fromFile(__DIR__ . '/resources/corrupt.jpg');
+    }
+
+    public function testSaveBadFileName(): void
+    {
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
+
+        $this->expectException(\InvalidArgumentException::class);
+        $this->expectExceptionMessage('File name does not contain a . extension (example.jpg)');
+
+        $image->save(self::$output_dir . '/foobar');
+    }
+
+    public function testUnsupportedFileName(): void
+    {
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
+
+        $this->expectException(UnsupportedExtensionException::class);
+
+        $image->save(self::$output_dir . '/foo.test');
+    }
+
+    /**
+     * @dataProvider orientationProvider
+     */
+    public function testOrientation(int $expected, string $filename): void
+    {
+        $image = Image::fromFile($filename);
+
+        self::assertSame($expected, $image->orientation());
+    }
+
+    /**
+     * @dataProvider fitProvider
+     */
+    public function testFit(array $expected_rect, bool $exact, string $filename): void
+    {
+        $image = Image::fromFile($filename);
+
+        self::assertSame($expected_rect, $image->fit(100, 100, $exact)->rect());
+    }
+
     public function testGeneric(): void
     {
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image2.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
 
         self::assertSame(700, $image->width());
         self::assertSame(875, $image->height());
-        self::assertSame(0.0, $image->duration());
+        self::assertSame([0, 0, 700, 875], $image->rect());
         self::assertSame([124, 192, 217, 0], $image->color(0, 0));
+        self::assertSame([0, 0, 100, 100], $image->resize(100, 100)->rect());
+        self::assertSame([0, 0, 100, 100], $image->crop([10, 10, 100, 100])->rect());
         self::assertIsString($image->data('.png'));
 
         $sampled = $image->sampleTo([0, 0, 50, 50]);
@@ -39,19 +159,21 @@ class StaticImageTest extends TestCase
         self::assertSame(50, $sampled->height());
         self::assertSame([124, 192, 217, 0], $sampled->color(0, 0));
 
-        $cut = $image->slice(0, 1);
-
-        self::assertSame(0.0, $cut->duration());
-        self::assertSame([124, 192, 217, 0], $cut->color(0, 0));
-
         $lower_quality = $image->quality(0.5);
 
         self::assertSame([124, 192, 217, 0], $lower_quality->color(0, 0));
+
+        $fs = new Filesystem();
+        $fs->mkdir(self::$output_dir);
+
+        $image->save(self::$output_dir . '/out.png');
+
+        self::assertFileExists(self::$output_dir . '/out.png');
     }
 
     public function testSavePng(): void
     {
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image2.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
 
         $fs = new Filesystem();
         $fs->mkdir(self::$output_dir);
@@ -66,7 +188,7 @@ class StaticImageTest extends TestCase
             $this->markTestSkipped('No webp support from current PHP version.');
         }
 
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image2.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
 
         $fs = new Filesystem();
         $fs->mkdir(self::$output_dir);
@@ -77,7 +199,7 @@ class StaticImageTest extends TestCase
 
     public function testSampleToError(): void
     {
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image2.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
 
         $this->expectException(ImageException::class);
         $image->sampleTo([0, 0, 1, 1], [0, 0, 0, 0]);
@@ -88,7 +210,7 @@ class StaticImageTest extends TestCase
      */
     public function testQualityInvalid(int $quality): void
     {
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image2.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image2.png');
 
         $this->expectException(\InvalidArgumentException::class);
         $image->quality($quality);
@@ -105,7 +227,7 @@ class StaticImageTest extends TestCase
 
     public function testTransparency(): void
     {
-        $image = AbstractImage::fromFile(__DIR__ . '/resources/image8.png');
+        $image = Image::fromFile(__DIR__ . '/resources/image8.png');
 
         self::assertSame(127, $image->color(0, 0)[3]);
         self::assertSame(0, $image->color(50, 50)[3]);
@@ -124,15 +246,6 @@ class StaticImageTest extends TestCase
 
             self::assertSame(127, imagecolorsforindex($res, imagecolorat($res, 0, 0))['alpha']);
             self::assertSame(0, imagecolorsforindex($res, imagecolorat($res, 25, 25))['alpha']);
-        } finally {
-            imagedestroy($res);
-        }
-
-        try {
-            $res = imagecreatefromstring($image->slice(0)->data('.png'));
-
-            self::assertSame(127, imagecolorsforindex($res, imagecolorat($res, 0, 0))['alpha']);
-            self::assertSame(0, imagecolorsforindex($res, imagecolorat($res, 50, 50))['alpha']);
         } finally {
             imagedestroy($res);
         }
